+++ components/cards/CardImageHeader.js
@@ -2,22 +2,36 @@ import React, { PropTypes } from 'react';
 
 
 export default function CardImageHeader(props) {
-  const { icon, iconClass, nav, title, subtitle } = props;
+  const { icon, iconClass, nav, title, subtitle, tags } = props;
 
   let img;
   if (icon) {
-    img = (<img
-      className={`CardImageHeader-icon float-sm-left ${iconClass}`}
-      src={icon}
-      alt="Client thumbnail"
-    />);
+    img = (
+      <img
+        className={`CardImageHeader-icon float-sm-left ${iconClass}`}
+        src={icon}
+        alt="Client thumbnail"
+      />
+    );
+  }
+
+  let headerTags;
+  if (tags) {
+    headerTags = (
+      <div className="CardImageHeader-tags float-sm-left">
+        {tags.map(tag => <div key={tag} className="CardImageHeader-tag">{tag}</div>)}
+      </div>
+    );
   }
 
   let headerTitle;
   if (title) {
     headerTitle = (
       <div className="float-sm-left">
-        <h2 className="CardImageHeader-title">{title}</h2>
+        <div className="clearfix">
+          <h2 className="CardImageHeader-title float-sm-left">{title}</h2>
+          {headerTags}
+        </div>
         {subtitle ? <div><small className="text-muted">{subtitle}</small></div> : null}
       </div>
     );
@@ -43,9 +57,11 @@ CardImageHeader.propTypes = {
   title: PropTypes.string,
   subtitle: PropTypes.object,
   nav: PropTypes.node,
+  tags: PropTypes.arrayOf(PropTypes.string).isRequired,
 };
 
 CardImageHeader.defaultProps = {
   iconClass: '',
   title: '',
+  tags: [],
 };