+++ src/linodes/linode/settings/advanced/layouts/IndexPage.js
@@ -1,7 +1,6 @@
 import { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 
-import { setError } from '~/actions/errors';
 import { linodes, kernels, distributions } from '~/api';
 import { getObjectByLabelLazily } from '~/api/util';
 
@@ -12,28 +11,24 @@ export class IndexPage extends Component {
   static async preload({ dispatch, getState }, { linodeLabel }) {
     const { id } = await dispatch(getObjectByLabelLazily('linodes', linodeLabel));
 
-    try {
-      const requests = [
-        linodes.disks.all([id]),
-        kernels.page(0),
-      ];
+    const requests = [
+      linodes.disks.all([id]),
+      kernels.page(0),
+    ];
 
-      await Promise.all(requests.map(r => dispatch(r)));
+    await Promise.all(requests.map(r => dispatch(r)));
 
-      // Fetch distributions without waiting
-      if (!getState().api.distributions.ids.length) {
-        // Only needed from the add disk modal which isn't shown immediately.
-        dispatch(distributions.all());
-      }
+    // Fetch distributions without waiting
+    if (!getState().api.distributions.ids.length) {
+      // Only needed from the add disk modal which isn't shown immediately.
+      dispatch(distributions.all());
+    }
 
-      const { totalPages } = getState().api.kernels;
+    const { totalPages } = getState().api.kernels;
 
-      // Fetch rest of kernel pages without waiting.
-      for (let i = 1; i < totalPages; i++) {
-        dispatch(kernels.page(i));
-      }
-    } catch (e) {
-      dispatch(setError(e));
+    // Fetch rest of kernel pages without waiting.
+    for (let i = 1; i < totalPages; i++) {
+      dispatch(kernels.page(i));
     }
   }